Bootstrap, plants, and style work
Add the bootstrap source files through client-side library in order to begin customising bootstrap objects. Add SASS files and a route for partial SASS files to be combined. Organise SASS and css files.

Add a plant section as a placeholder.

Work on styles for the database, index, and layout views.